refactor: split functions into separate files
