+++ src/decodeBase64.ts
@@ -1,0 +1,3 @@
+export default function decodeBase64(base64: string): Uint8Array {
+  return new Uint8Array(Buffer.from(base64, 'base64'));
+}
