+++ src/encodeUint8Array.ts
@@ -1,0 +1,3 @@
+export default function encodeUint8Array(uint8Array: Uint8Array): string {
+  return Buffer.from(uint8Array).toString('base64');
+}
